Rename test helper module
So it's not taking the canonical name for the in-file unit test
module. Also, it's a better name for it.
